fix memory leak + rename + restructure
