+++ src/index.tsx
@@ -11,8 +11,8 @@ import { UserProvider } from './context/userContext/UserContext';
 import { FiltersProvider } from './context/filtersContext/FiltersContext';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import { ModalHookProvider } from './context/modalContext/ModalContext';
-import { LazyRoute } from './components/Routers/LazyRoute';
-import { AuthorizationRoute } from './Me/AuthorizationRoute';
+import { LazyRoute } from './CustomRoutes/LazyRoute';
+import { AuthorizationRoute } from './CustomRoutes/AuthorizedRoute';
 
 const PageNotFound = lazy(() =>
   import(/* webpackChunkName: "PageNotFound" */ './PageNotFound')
